pyproject.toml
[tool.poetry]
name = "walletapi"
version = "0.1.0"
description = ""
authors = ["krittamet-rrt <krittamet.rrt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.21"
pg8000 = "^1.31.2"
fastapi = "^0.112.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"
passlib = "^1.7.4"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
